forge: remove AdminInfo type
Remove the AdminInfo type and just use ChainInfo instead.
